requirements.txt
numpy==1.26.2
torch==2.1.2
scipy==1.11.4
sympy==1.12
tqdm==4.65.0
wandb==0.16.0 
pyyaml==6.0.1
h5py==3.10.0
matplotlib==3.8.3
pytorch-lightning==2.2.0
gymnasium==0.29.1
joblib==1.4.2
python-box==7.2.0
scikit-learn==1.5.2